[back] refactor : TestRequest, TestResponse 파일 이동 및 코드 주석 처리 제거
