+++ back/src/main/java/com/example/cpsplatform/ai/service/AiApiService.java
@@ -1,8 +1,5 @@
 package com.example.cpsplatform.ai.service;
 
-import com.example.cpsplatform.TestRequest;
-import com.example.cpsplatform.TestResponse;
-
 
 public interface AiApiService {
