+++ back/src/main/java/com/example/cpsplatform/ai/service/TestRequest.java
@@ -1,0 +1,10 @@
+package com.example.cpsplatform.ai.service;
+
+import lombok.Getter;
+
+@Getter
+public class TestRequest {
+
+    private String request;
+
+}
